+++ pp/routing/__init__.py
@@ -8,6 +8,7 @@ from pp.routing.add_electrical_pads_top import add_electrical_pads_top
 from pp.routing.add_fiber_array import add_fiber_array
 from pp.routing.add_fiber_single import add_fiber_single
 from pp.routing.connect import connect_strip, connect_strip_way_points
+from pp.routing.connect import connect_elec_waypoints
 from pp.routing.connect_bundle import connect_bundle
 from pp.routing.connect_bundle import connect_bundle_path_length_match
 from pp.routing.connect_bundle import link_electrical_ports
@@ -16,6 +17,10 @@ from pp.routing.manhattan import round_corners, route_manhattan
 from pp.routing.repackage import package_optical2x2
 from pp.routing.route_fiber_single import route_fiber_single
 from pp.routing.route_ports_to_side import route_elec_ports_to_side, route_ports_to_side
+
+route_factory = dict(
+    optical=connect_strip_way_points, electrical=connect_elec_waypoints
+)
 
 __all__ = [
     "add_electrical_pads",
@@ -37,7 +42,9 @@ __all__ = [
     "route_manhattan",
     "route_ports_to_side",
     "route_ports_to_side",
+    "route_factory",
 ]
+
 
 if __name__ == "__main__":
     print(__all__)